requirements.txt
requests==2.21.0
instabot==0.34.0
pytz==2019.1
python_dateutil==2.8.0
python-dotenv==0.10.1
